requirements.txt
digitalai-release-sdk == 24.1.0
requests >= 2.28.1
python-jenkins == 1.8.2
